+++ inc/Constants.php
@@ -19,5 +19,5 @@ class Constants extends LibConstants
 
     public const DEFAULT_BASE_FILE_PATH = __DIR__ . '/../../../../wp-content/uploads/crowdsec/';
     public const STANDALONE_CONFIG_PATH = __DIR__ . '/standalone-settings.php';
-    public const VERSION = 'v2.5.2';
+    public const VERSION = 'v2.6.0';
 }